dev_requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
more-itertools==4.2.0
paho-mqtt==1.3.1
pika==0.12.0
pluggy==0.6.0
py==1.5.3
pytest==3.6.1
six==1.11.0
